Rename skill adapter. Updated OnError Upated nuget packages to released versions.
