+++ experimental/skills/DialogToDialog/DialogSkillBot/DialogSkillBot.csproj
@@ -10,9 +10,9 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Microsoft.Recognizers.Text.DataTypes.TimexExpression" Version="1.1.5" />
-    <PackageReference Include="Microsoft.Bot.Builder.AI.Luis" Version="4.7.0-rc1" />
-    <PackageReference Include="Microsoft.Bot.Builder.Dialogs" Version="4.7.0-rc1" />
-    <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="4.7.0-rc1" />
+    <PackageReference Include="Microsoft.Bot.Builder.AI.Luis" Version="4.7.0" />
+    <PackageReference Include="Microsoft.Bot.Builder.Dialogs" Version="4.7.0" />
+    <PackageReference Include="Microsoft.Bot.Builder.Integration.AspNet.Core" Version="4.7.0" />
   </ItemGroup>
 
   <ItemGroup>
